requirements.txt
numpy>=1.24.0
scikit-learn>=1.3.0
flask>=2.3.0
mysql-connector-python>=9.0.0
matplotlib>=3.8.0
seaborn>=0.13.2
pandas>=2.1.0

